+++ datastore/couchDB/couchDatastoreImpl_test.go
@@ -115,4 +115,5 @@ func RunTenantServiceDatastoreTests(t *testing.T) {
 
 	tester.RunBrandingCRUD(t)
 	tester.RunLocaleCRUD(t)
+	tester.RunTenantMetadataConfigCRUD(t)
 }